metric-agent/requirements.txt
grpcio>=1.74.0
grpcio-tools>=1.74.0
psutil>=7.0.0
